% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVK-package.R
\docType{package}
\name{CVK-package}
\alias{CVK}
\alias{CVK-package}
\title{CVK: Simple Repeated Holdout Cross-Validation for rpart}
\description{
Provides a minimal cross_validate() helper that repeats random train/validation splits for rpart models.
}
\author{
\strong{Maintainer}: Kashish Vaish \email{kashishvaish@gmail.com}

}
\keyword{internal}
